Refactor stellar cartography structure for maintainability.
Unify cartography geometry, complete per-analytic map merge registry, decompose MapGraph, centralize Python layer constants, add cross-layer ion voltage contract tests, split BFF cartography types, and bundle map UI props with a wormhole interaction provider.
