+++ packages/api/api/concepts/stellar_cartography/__init__.py
@@ -1,5 +1,1 @@
-"""Stellar Cartography turn-scoped sampling (map hover tooltips)."""
-
-from api.concepts.stellar_cartography.sample_at import sample_at
-
-__all__ = ["sample_at"]
+"""Stellar Cartography map overlays and cell sampling."""
