+++ packages/api/api/concepts/stellar_cartography/layers.py
@@ -1,0 +1,15 @@
+"""Stellar Cartography overlay layer wire ids (shared by map analytic and sampling)."""
+
+LAYER_NEBULAE = "nebulae"
+LAYER_ION_STORMS = "ion-storms"
+LAYER_STAR_CLUSTERS = "star-clusters"
+LAYER_BLACK_HOLES = "black-holes"
+LAYER_DEBRIS_DISKS = "debris-disks"
+LAYER_WORMHOLES = "wormholes"
+
+PAINT_ORDER = (
+    LAYER_NEBULAE,
+    LAYER_ION_STORMS,
+    LAYER_STAR_CLUSTERS,
+    LAYER_BLACK_HOLES,
+)
