+++ packages/frontend/src/components/MapGraph.tsx
@@ -1,58 +1,26 @@
 import {
   useCallback,
   useEffect,
-  useLayoutEffect,
   useMemo,
-  useRef,
   useState,
 } from 'react'
-import {
-  ReactFlow,
-  Panel,
-  useReactFlow,
-  useStore,
-  useStoreApi,
-} from '@xyflow/react'
+import { ReactFlow } from '@xyflow/react'
 import '@xyflow/react/dist/style.css'
-import type { CombinedMapData, RouteMapWaypoint, AnalyticShellScope } from '../api/bff'
+import type { CombinedMapData } from '../api/bff'
 import { StellarCartographyHoverPanel } from '../analytics/stellar-cartography/StellarCartographyHoverPanel'
-import type {
-  CartographyLayerVisibility,
-  StellarCartographySettingsGates,
-} from '../analytics/stellar-cartography/layers'
 import {
   filterWormholeEdgesForDisplayMode,
   type WormholeDisplayMode,
 } from '../analytics/stellar-cartography/wormholeDisplayMode'
-import { WORMHOLE_RECENTER_PULSE_MS } from '../lib/stellarCartographyTheme'
 import {
   buildWormholeEndpointHoverIndex,
-  wormholeMapCellKey,
 } from '../lib/wormholeEndpointHover'
-import {
-  buildPlanetSpatialGrid,
-  findClosestPlanetWithinRadius,
-  flowCenterToPlanet,
-  type PlanetSpatialGrid,
-} from '../lib/planetSpatialGrid'
-import { cn } from '../lib/utils'
-import { PlanetMapLabel } from './PlanetMapLabel'
+import { buildPlanetSpatialGrid } from '../lib/planetSpatialGrid'
 import {
   DEFAULT_PLANET_LABEL_OPTIONS,
-  planetLabelOptionsShowAnyLabel,
   type PlanetLabelOptions,
 } from './planetMapLabelModel'
-import {
-  buildWarpWellOverlayPaneLines,
-  WARP_WELL_OVERLAY_ZOOM_THRESHOLD,
-} from '../lib/warpWellOverlay'
-import {
-  CELL_CENTER_OFFSET,
-  clientToFlowPosition,
-  flowCenterFromMapNode,
-  gameMapYToFlowCenterY,
-  safeZoomScale,
-} from './map-graph/geometry'
+import { clientToFlowPosition } from './map-graph/geometry'
 import { nodeTypes, toFlowNodes } from './map-graph/nodes'
 import { edgeTypes, toEdges } from './map-graph/edges'
 import {
@@ -60,655 +28,21 @@ import {
   collectWormholeEndpoints,
 } from './map-graph/StellarCartographyOverlayPane'
 import {
-  WORMHOLE_LINE_REVEAL_CLEAR_MS,
-  WormholeHoverContext,
-  WormholeLineRevealContext,
-  WormholeRecenterPulseContext,
-  type WormholeLineRevealApi,
-  type WormholeRecenterPulseTarget,
+  WormholeInteractionProvider,
+  useWormholeInteractionState,
 } from './map-graph/stellarCartographyWormholeInteraction'
-
-/** Fixed pixel size of the planet dot on screen (independent of zoom). */
-const DOT_PIXELS = 4
-/** On-screen size of a multi-hop route intermediate marker (smaller and quieter than planet dots). */
-const ROUTE_WAYPOINT_CROSS_PX = 5
-/** Mouse distance from dot center (px) at which the planet label is shown. */
-const PLANET_LABEL_HOVER_RADIUS_PX = 14
-
-/** Fraction of display area to leave as blank margin on each side when fitting initial view (0.1 = 10%). */
-const INITIAL_FIT_MARGIN = 0.1
-
-/**
- * Computes initial viewport so the bounding rectangle of all node centers is
- * centered with a 10% margin (whichever dimension is most constrained).
- * Runs once when the pane has size and nodes are present. Calls onInitialFitDone when done (or when no fit will run).
- */
-function InitialViewportFit({
-  nodes,
-  onInitialFitDone,
-  onMapZoomChange,
-}: {
-  nodes: CombinedMapData['nodes']
-  onInitialFitDone: () => void
-  onMapZoomChange: (zoom: number) => void
-}) {
-  const { setViewport } = useReactFlow()
-  const domNode = useStore((s) => s.domNode ?? null)
-  const [size, setSize] = useState({ width: 0, height: 0 })
-  const hasFittedRef = useRef(false)
-  const doneCalledRef = useRef(false)
-
-  const callDoneOnce = useCallback(() => {
-    if (doneCalledRef.current) return
-    doneCalledRef.current = true
-    onInitialFitDone()
-  }, [onInitialFitDone])
-
-  useEffect(() => {
-    if (!domNode) return
-    let raf = 0
-    const ro = new ResizeObserver((entries) => {
-      const { width, height } = entries[0]?.contentRect ?? { width: 0, height: 0 }
-      cancelAnimationFrame(raf)
-      raf = requestAnimationFrame(() => setSize({ width, height }))
-    })
-    ro.observe(domNode)
-    return () => {
-      cancelAnimationFrame(raf)
-      ro.disconnect()
-    }
-  }, [domNode])
-
-  useEffect(() => {
-    if (nodes.length === 0) {
-      callDoneOnce()
-      return
-    }
-    if (size.width <= 0 || size.height <= 0 || hasFittedRef.current) return
-    const xs = nodes.map((n) => Number(n.x)).filter(Number.isFinite)
-    const ys = nodes.map((n) => Number(n.y)).filter(Number.isFinite)
-    if (xs.length === 0 || ys.length === 0) {
-      callDoneOnce()
-      return
-    }
-    const minFx = Math.min(...xs) + CELL_CENTER_OFFSET
-    const maxFx = Math.max(...xs) + CELL_CENTER_OFFSET
-    const flowCentersY = ys.map((py) => gameMapYToFlowCenterY(py))
-    const minFy = Math.min(...flowCentersY)
-    const maxFy = Math.max(...flowCentersY)
-    const contentWidth = Math.max(maxFx - minFx, 1)
-    const contentHeight = Math.max(maxFy - minFy, 1)
-    const centerX = (minFx + maxFx) / 2
-    const centerY = (minFy + maxFy) / 2
-    const usableW = size.width * (1 - 2 * INITIAL_FIT_MARGIN)
-    const usableH = size.height * (1 - 2 * INITIAL_FIT_MARGIN)
-    const scaleW = usableW / contentWidth
-    const scaleH = usableH / contentHeight
-    const zoom = Math.min(40, Math.max(0.2, Math.min(scaleW, scaleH)))
-    const x = size.width / 2 - centerX * zoom
-    const y = size.height / 2 - centerY * zoom
-    hasFittedRef.current = true
-    setViewport({ x, y, zoom })
-    onMapZoomChange(zoom)
-    callDoneOnce()
-  }, [nodes, size, setViewport, callDoneOnce, onMapZoomChange])
-
-  return null
-}
-
-/**
- * Tracks mouse over the flow viewport and shows position in flow coordinates.
- * Attaches listeners to the store's domNode so we receive events regardless of
- * whether ReactFlow forwards onMouseMove. Must be rendered inside ReactFlow.
- */
-function FlowCoordinateReadout() {
-  const domNode = useStore((s) => s.domNode ?? null)
-  const transform = useStore((s) => s.transform)
-  const [clientPos, setClientPos] = useState<{ x: number; y: number } | null>(null)
-
-  useEffect(() => {
-    const el = domNode
-    if (!el) return
-    const onMove = (e: MouseEvent) => setClientPos({ x: e.clientX, y: e.clientY })
-    const onLeave = () => setClientPos(null)
-    el.addEventListener('mousemove', onMove)
-    el.addEventListener('mouseleave', onLeave)
-    return () => {
-      el.removeEventListener('mousemove', onMove)
-      el.removeEventListener('mouseleave', onLeave)
-    }
-  }, [domNode])
-
-  const flow = clientPos
-    ? clientToFlowPosition(clientPos.x, clientPos.y, domNode, transform)
-    : null
-  const scale = typeof transform?.[2] === 'number' && Number.isFinite(transform[2]) ? transform[2] : null
-
-  const content =
-    clientPos == null ? (
-      scale != null ? <>zoom: {scale.toFixed(2)}</> : '—'
-    ) : flow != null ? (
-      <>
-        x: {Math.floor(flow.x)} y: {Math.floor(-flow.y)} zoom: {scale != null ? scale.toFixed(2) : '—'}
-      </>
-    ) : (
-      <>client: {Math.round(clientPos.x)}, {Math.round(clientPos.y)} zoom: {scale != null ? scale.toFixed(2) : '—'}</>
-    )
-
-  return (
-    <Panel position="bottom-left" className="rounded bg-black/80 px-2 py-1 font-mono text-xs text-gray-300">
-      {content}
-    </Panel>
-  )
-}
-
-/** Show grid when zoom >= this (pixels per flow unit). */
-const GRID_ZOOM_THRESHOLD = 15
-
-/** Light grey at 30% opacity so the warp-well overlay reads stronger when lines coincide. */
-const GRID_STROKE = 'rgba(107, 114, 128, 0.3)'
-
-/** Slightly warmer than the coordinate grid so both remain distinguishable. */
-const WARP_WELL_STROKE = '#78716c'
-
-/**
- * Coordinate grid overlay when zoomed in. Drawn in pixel space so lines stay 1px at any zoom;
- * flow positions converted to pane pixels via pane = flow * scale + translation.
- */
-function CoordinateGridOverlay() {
-  const domNode = useStore((s) => s.domNode ?? null)
-  const transform = useStore((s) => s.transform)
-  const [size, setSize] = useState({ width: 0, height: 0 })
-
-  useEffect(() => {
-    if (!domNode) return
-    let raf = 0
-    const ro = new ResizeObserver((entries) => {
-      const { width, height } = entries[0]?.contentRect ?? { width: 0, height: 0 }
-      cancelAnimationFrame(raf)
-      raf = requestAnimationFrame(() => setSize({ width, height }))
-    })
-    ro.observe(domNode)
-    return () => {
-      cancelAnimationFrame(raf)
-      ro.disconnect()
-    }
-  }, [domNode])
-
-  if (!transform || size.width <= 0 || size.height <= 0) return null
-  const [tx, ty, rawScale] = transform
-  const scale = safeZoomScale(rawScale)
-  if (scale < GRID_ZOOM_THRESHOLD) return null
-
-  const { width, height } = size
-  const flowXMin = -tx / scale
-  const flowXMax = (width - tx) / scale
-  const flowYMin = -ty / scale
-  const flowYMax = (height - ty) / scale
-
-  const xFrom = Math.floor(flowXMin)
-  const xTo = Math.ceil(flowXMax)
-  const yFrom = Math.floor(flowYMin)
-  const yTo = Math.ceil(flowYMax)
-
-  const verticals = Array.from({ length: xTo - xFrom + 1 }, (_, i) => {
-    const flowX = xFrom + i
-    const paneX = flowX * scale + tx
-    return { key: `v${flowX}`, x: paneX }
-  })
-  const horizontals = Array.from({ length: yTo - yFrom + 1 }, (_, i) => {
-    const flowY = yFrom + i
-    const paneY = flowY * scale + ty
-    return { key: `h${flowY}`, y: paneY }
-  })
-
-  return (
-    <div
-      className="pointer-events-none absolute inset-0 z-[5]"
-      aria-hidden
-    >
-      <svg
-        className="h-full w-full"
-        viewBox={`0 0 ${width} ${height}`}
-        preserveAspectRatio="none"
-      >
-        <g stroke={GRID_STROKE} strokeWidth={1}>
-          {verticals.map(({ key, x }) => (
-            <line key={key} x1={x} y1={0} x2={x} y2={height} />
-          ))}
-          {horizontals.map(({ key, y }) => (
-            <line key={key} x1={0} y1={y} x2={width} y2={y} />
-          ))}
-        </g>
-      </svg>
-    </div>
-  )
-}
-
-function NormalWarpWellOutlinesOverlay({ mapNodes }: { mapNodes: CombinedMapData['nodes'] }) {
-  const domNode = useStore((s) => s.domNode ?? null)
-  const transform = useStore((s) => s.transform)
-  const [size, setSize] = useState({ width: 0, height: 0 })
-
-  useEffect(() => {
-    if (!domNode) return
-    let raf = 0
-    const ro = new ResizeObserver((entries) => {
-      const { width, height } = entries[0]?.contentRect ?? { width: 0, height: 0 }
-      cancelAnimationFrame(raf)
-      raf = requestAnimationFrame(() => setSize({ width, height }))
-    })
-    ro.observe(domNode)
-    return () => {
-      cancelAnimationFrame(raf)
-      ro.disconnect()
-    }
-  }, [domNode])
-
-  if (!transform || size.width <= 0 || size.height <= 0) return null
-  const [tx, ty, rawScale] = transform
-  const scale = safeZoomScale(rawScale)
-  if (scale < WARP_WELL_OVERLAY_ZOOM_THRESHOLD) return null
-
-  const { width, height } = size
-  const lines = buildWarpWellOverlayPaneLines(
-    mapNodes,
-    { width, height, tx, ty, scale },
-    WARP_WELL_OVERLAY_ZOOM_THRESHOLD
-  )
-
-  if (lines.length === 0) return null
-
-  return (
-    <div className="pointer-events-none absolute inset-0 z-[5]" aria-hidden>
-      <svg className="h-full w-full" viewBox={`0 0 ${width} ${height}`} preserveAspectRatio="none">
-        <g stroke={WARP_WELL_STROKE} strokeWidth={1}>
-          {lines.map(({ key, x1, y1, x2, y2 }) => (
-            <line key={key} x1={x1} y1={y1} x2={x2} y2={y2} />
-          ))}
-        </g>
-      </svg>
-    </div>
-  )
-}
-
-/**
- * in size regardless of zoom. Uses same flow->pane conversion as the grid.
- */
-const HOVER_CLIENT_MOVE_EPS_PX = 0.5
-
-/** Label text uses map payload (planet name, etc.). React Flow's internal node store does not reliably retain custom `data` fields. */
-type MapNodeLabelSource = {
-  planet?: Record<string, unknown>
-  ownerName?: string | null
-  mapX: number
-  mapY: number
-}
-
-function buildLabelSourceByNodeId(nodes: CombinedMapData['nodes']): Map<string, MapNodeLabelSource> {
-  const m = new Map<string, MapNodeLabelSource>()
-  for (const n of nodes) {
-    const payload: MapNodeLabelSource = {
-      planet: n.planet,
-      ownerName: n.ownerName ?? null,
-      mapX: Number(n.x),
-      mapY: Number(n.y),
-    }
-    m.set(n.id, payload)
-  }
-  return m
-}
-
-function FixedSizeDotsOverlay({
-  planetGrid,
-  planetLabelOptions,
-  labelSourceByNodeId,
-  mapNodes,
-  routeWaypoints,
-  waypointGrid,
-  onPlanetLabelHoverActiveChange,
-}: {
-  planetGrid: PlanetSpatialGrid | null
-  planetLabelOptions: PlanetLabelOptions
-  labelSourceByNodeId: Map<string, MapNodeLabelSource>
-  mapNodes: CombinedMapData['nodes']
-  routeWaypoints: readonly RouteMapWaypoint[]
-  /** Sub-linear hover: same map-cell + radius model as :func:`buildPlanetSpatialGrid` for planets. */
-  waypointGrid: PlanetSpatialGrid | null
-  onPlanetLabelHoverActiveChange?: (active: boolean) => void
-}) {
-  const domNode = useStore((s) => s.domNode ?? null)
-  const transform = useStore((s) => s.transform)
-  const [size, setSize] = useState({ width: 0, height: 0 })
-  const [hoveredNodeId, setHoveredNodeId] = useState<string | null>(null)
-  const [hoveredWaypointId, setHoveredWaypointId] = useState<string | null>(null)
-  const [pinnedNodeId, setPinnedNodeId] = useState<string | null>(null)
-  const hoverRafRef = useRef<number | null>(null)
-  const pendingClientRef = useRef<{ x: number; y: number } | null>(null)
-  const lastProcessedClientRef = useRef<{ x: number; y: number } | null>(null)
-  const transformRef = useRef(transform)
-  const pinnedNodeIdRef = useRef<string | null>(null)
-  useLayoutEffect(() => {
-    transformRef.current = transform
-  }, [transform])
-  useLayoutEffect(() => {
-    pinnedNodeIdRef.current = pinnedNodeId
-  }, [pinnedNodeId])
-
-  const showAnyLabelOption = planetLabelOptionsShowAnyLabel(planetLabelOptions)
-
-  const planetLabelHoverActive =
-    pinnedNodeId != null || (showAnyLabelOption && hoveredNodeId != null)
-
-  useEffect(() => {
-    onPlanetLabelHoverActiveChange?.(planetLabelHoverActive)
-  }, [onPlanetLabelHoverActiveChange, planetLabelHoverActive])
-
-  const mapNodeIdsKey = useMemo(() => mapNodes.map((n) => n.id).join('\0'), [mapNodes])
-
-  useEffect(() => {
-    setPinnedNodeId(null)
-  }, [mapNodeIdsKey])
-
-  useEffect(() => {
-    if (!showAnyLabelOption) {
-      setPinnedNodeId(null)
-    }
-  }, [showAnyLabelOption])
-
-  useEffect(() => {
-    if (pinnedNodeId == null) return
-    const onKeyDown = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') setPinnedNodeId(null)
-    }
-    window.addEventListener('keydown', onKeyDown)
-    return () => window.removeEventListener('keydown', onKeyDown)
-  }, [pinnedNodeId])
-
-  useEffect(() => {
-    if (pinnedNodeId != null) {
-      setHoveredNodeId(null)
-      setHoveredWaypointId(null)
-    }
-  }, [pinnedNodeId])
-
-  const routeWaypointIdSet = useMemo(
-    () => new Set(routeWaypoints.map((w) => w.id)),
-    [routeWaypoints]
-  )
-  const hoveredWaypointInList =
-    hoveredWaypointId != null && routeWaypointIdSet.has(hoveredWaypointId)
-  const hoveredWaypointIdForLabel =
-    pinnedNodeId == null && hoveredWaypointInList ? hoveredWaypointId : null
-
-  useEffect(() => {
-    if (!domNode) return
-    let raf = 0
-    const ro = new ResizeObserver((entries) => {
-      const { width, height } = entries[0]?.contentRect ?? { width: 0, height: 0 }
-      cancelAnimationFrame(raf)
-      raf = requestAnimationFrame(() => setSize({ width, height }))
-    })
-    ro.observe(domNode)
-    return () => {
-      cancelAnimationFrame(raf)
-      ro.disconnect()
-    }
-  }, [domNode])
-
-  useEffect(() => {
-    const el = domNode
-    if (!el || size.width <= 0 || size.height <= 0) return
-
-    if (!planetGrid && !waypointGrid) {
-      return
-    }
-
-    const runHitTest = (clientX: number, clientY: number) => {
-      if (pinnedNodeIdRef.current != null) return
-      const t = transformRef.current
-      if (!t) {
-        setHoveredNodeId(null)
-        setHoveredWaypointId(null)
-        return
-      }
-      const paneRect = el.getBoundingClientRect()
-      const flow = clientToFlowPosition(clientX, clientY, el, t, paneRect)
-      if (!flow) {
-        setHoveredNodeId(null)
-        setHoveredWaypointId(null)
-        return
-      }
-      const rawScale = t[2]
-      const scale = safeZoomScale(rawScale)
-      const radiusFlow = PLANET_LABEL_HOVER_RADIUS_PX / scale
-      if (planetGrid) {
-        const { px, py } = flowCenterToPlanet(flow.x, flow.y)
-        const closestId = findClosestPlanetWithinRadius(planetGrid, px, py, radiusFlow)
-        if (closestId != null) {
-          setHoveredNodeId(closestId)
-          setHoveredWaypointId(null)
-          return
-        }
-      }
-      setHoveredNodeId(null)
-      if (waypointGrid) {
-        const { px, py } = flowCenterToPlanet(flow.x, flow.y)
-        setHoveredWaypointId(findClosestPlanetWithinRadius(waypointGrid, px, py, radiusFlow))
-      } else {
-        setHoveredWaypointId(null)
-      }
-    }
-
-    const flushHover = () => {
-      const p = pendingClientRef.current
-      if (!p) return
-      const last = lastProcessedClientRef.current
-      if (
-        last &&
-        Math.abs(p.x - last.x) < HOVER_CLIENT_MOVE_EPS_PX &&
-        Math.abs(p.y - last.y) < HOVER_CLIENT_MOVE_EPS_PX
-      ) {
-        return
-      }
-      lastProcessedClientRef.current = { x: p.x, y: p.y }
-      runHitTest(p.x, p.y)
-    }
-
-    const onMove = (e: MouseEvent) => {
-      pendingClientRef.current = { x: e.clientX, y: e.clientY }
-      if (hoverRafRef.current != null) return
-      hoverRafRef.current = requestAnimationFrame(() => {
-        hoverRafRef.current = null
-        flushHover()
-      })
-    }
-    const onLeave = () => {
-      pendingClientRef.current = null
-      lastProcessedClientRef.current = null
-      if (hoverRafRef.current != null) {
-        cancelAnimationFrame(hoverRafRef.current)
-        hoverRafRef.current = null
-      }
-      setHoveredNodeId(null)
-      setHoveredWaypointId(null)
-    }
-    el.addEventListener('mousemove', onMove)
-    el.addEventListener('mouseleave', onLeave)
-    return () => {
-      if (hoverRafRef.current != null) cancelAnimationFrame(hoverRafRef.current)
-      hoverRafRef.current = null
-      el.removeEventListener('mousemove', onMove)
-      el.removeEventListener('mouseleave', onLeave)
-    }
-  }, [domNode, size.width, size.height, planetGrid, waypointGrid])
-
-  useEffect(() => {
-    const el = domNode
-    if (!el || !planetGrid) return
-
-    const onClick = (e: MouseEvent) => {
-      if (e.button !== 0) return
-      const t = transformRef.current
-      if (!t) return
-      const paneRect = el.getBoundingClientRect()
-      const flow = clientToFlowPosition(e.clientX, e.clientY, el, t, paneRect)
-      if (!flow) return
-      const rawScale = t[2]
-      const scale = safeZoomScale(rawScale)
-      const radiusPlanet = PLANET_LABEL_HOVER_RADIUS_PX / scale
-      const { px, py } = flowCenterToPlanet(flow.x, flow.y)
-      const closestId = findClosestPlanetWithinRadius(planetGrid, px, py, radiusPlanet)
-      if (closestId == null) {
-        if (pinnedNodeIdRef.current != null) {
-          setPinnedNodeId(null)
-        }
-        return
-      }
-      if (!showAnyLabelOption) {
-        if (pinnedNodeIdRef.current != null) {
-          setPinnedNodeId(null)
-        }
-        return
-      }
-      setPinnedNodeId((prev) => {
-        if (prev === closestId) return null
-        return closestId
-      })
-    }
-    el.addEventListener('click', onClick)
-    return () => el.removeEventListener('click', onClick)
-  }, [domNode, planetGrid, showAnyLabelOption])
-
-  if (!transform || size.width <= 0 || size.height <= 0) return null
-  const [tx, ty, rawScale] = transform
-  const scale = safeZoomScale(rawScale)
-  const hoveredForDisplay = planetGrid ? hoveredNodeId : null
-
-  const LABEL_OFFSET_X_PX = 9
-  const LABEL_OFFSET_Y_PX = -12
-
-  /**
-   * Dots and labels must not share one per-planet stacking group: later planets' dots were painting
-   * above earlier planets' labels (same z-index, DOM order), which looked like map bleed-through.
-   */
-  return (
-    <div
-      className="pointer-events-none absolute inset-0 z-[5]"
-      aria-hidden={pinnedNodeId == null ? true : undefined}
-    >
-      <div className="absolute inset-0" aria-hidden>
-        {routeWaypoints.map((w) => {
-          const { cx, cy } = flowCenterFromMapNode({ x: w.gx, y: w.gy })
-          const paneX = cx * scale + tx
-          const paneY = cy * scale + ty
-          const s = ROUTE_WAYPOINT_CROSS_PX
-          return (
-            <div
-              key={w.id}
-              className="absolute text-gray-500/75"
-              style={{
-                left: Math.round(paneX - s / 2),
-                top: Math.round(paneY - s / 2),
-                width: s,
-                height: s,
-              }}
-            >
-              <svg viewBox="0 0 8 8" className="h-full w-full" aria-hidden>
-                <line x1="1" y1="1" x2="7" y2="7" stroke="currentColor" strokeWidth="1.1" />
-                <line x1="7" y1="1" x2="1" y2="7" stroke="currentColor" strokeWidth="1.1" />
-              </svg>
-            </div>
-          )
-        })}
-      </div>
-      <div className="absolute inset-0" aria-hidden>
-        {mapNodes.map((mapNode) => {
-          const { cx, cy } = flowCenterFromMapNode(mapNode)
-          const paneX = cx * scale + tx
-          const paneY = cy * scale + ty
-          return (
-            <div
-              key={`dot-${mapNode.id}`}
-              className="absolute rounded-full bg-[#9ca3af]"
-              style={{
-                left: Math.round(paneX - DOT_PIXELS / 2),
-                top: Math.round(paneY - DOT_PIXELS / 2),
-                width: DOT_PIXELS,
-                height: DOT_PIXELS,
-              }}
-            />
-          )
-        })}
-      </div>
-      <div className="absolute inset-0 z-[1]">
-        {routeWaypoints.map((w) => {
-          if (hoveredWaypointIdForLabel !== w.id) return null
-          const { cx, cy } = flowCenterFromMapNode({ x: w.gx, y: w.gy })
-          const paneX = cx * scale + tx
-          const paneY = cy * scale + ty
-          return (
-            <div
-              key={`wpl-${w.id}`}
-              className="absolute font-mono text-gray-400"
-              style={{
-                left: Math.round(paneX - DOT_PIXELS / 2 + LABEL_OFFSET_X_PX),
-                top: Math.round(paneY - DOT_PIXELS / 2 + LABEL_OFFSET_Y_PX),
-                fontSize: 10,
-                backgroundColor: '#000000',
-                borderRadius: 6,
-                padding: '0 4px',
-              }}
-            >
-              {w.gx}, {w.gy}
-            </div>
-          )
-        })}
-        {mapNodes.map((mapNode) => {
-          const { cx, cy } = flowCenterFromMapNode(mapNode)
-          const paneX = cx * scale + tx
-          const paneY = cy * scale + ty
-          const labelSrc = labelSourceByNodeId.get(mapNode.id)
-          const coordX =
-            labelSrc != null && Number.isFinite(labelSrc.mapX) ? labelSrc.mapX : Number(mapNode.x)
-          const coordY =
-            labelSrc != null && Number.isFinite(labelSrc.mapY) ? labelSrc.mapY : Number(mapNode.y)
-          const isPinned = pinnedNodeId === mapNode.id
-          const showHoverLabel =
-            pinnedNodeId == null && showAnyLabelOption && hoveredForDisplay === mapNode.id
-          const showLabel = isPinned || showHoverLabel
-          if (!showLabel) return null
-          return (
-            <div
-              key={`label-${mapNode.id}`}
-              className={cn(
-                'absolute font-mono text-gray-300',
-                isPinned && 'pointer-events-auto z-[2]'
-              )}
-              style={{
-                left: Math.round(paneX - DOT_PIXELS / 2 + LABEL_OFFSET_X_PX),
-                top: Math.round(paneY - DOT_PIXELS / 2 + LABEL_OFFSET_Y_PX),
-                fontSize: 10,
-                backgroundColor: '#000000',
-                borderRadius: 6,
-              }}
-              onClick={isPinned ? (e) => e.stopPropagation() : undefined}
-            >
-              <PlanetMapLabel
-                options={planetLabelOptions}
-                nodeId={mapNode.id}
-                planet={labelSrc?.planet}
-                ownerName={labelSrc?.ownerName}
-                planetX={coordX}
-                planetY={coordY}
-              />
-            </div>
-          )
-        })}
-      </div>
-    </div>
-  )
-}
+import type { StellarCartographyMapUi } from './map-graph/stellarCartographyMapUi'
+import {
+  buildLabelSourceByNodeId,
+  FixedSizeDotsOverlay,
+} from './map-graph/FixedSizeDotsOverlay'
+import { CoordinateGridOverlay, FlowCoordinateReadout } from './map-graph/coordinateGrid'
+import { NormalWarpWellOutlinesOverlay } from './map-graph/NormalWarpWellOutlinesOverlay'
+import {
+  InitialViewportFit,
+  SliderZoomControl,
+  ViewportZoomSync,
+} from './map-graph/viewportControls'
 
 type MapGraphProps = {
   data: CombinedMapData
@@ -717,88 +51,13 @@ type MapGraphProps = {
   /** Called once so the header slider can drive zoom (same as scroll wheel). */
   onSetZoomReady: (setZoom: (zoom: number) => void) => void
   planetLabelOptions?: PlanetLabelOptions
-  analyticScope?: AnalyticShellScope | null
-  stellarCartographySampleEnabled?: boolean
-  cartographyLayerVisibility?: CartographyLayerVisibility
-  cartographySettingsGates?: StellarCartographySettingsGates
-  wormholeDisplayMode?: WormholeDisplayMode
-}
-
-/** Mirrors React Flow zoom to the app (wheel, pinch, initial fit, slider). */
-function ViewportZoomSync({ onMapZoomChange }: { onMapZoomChange: (z: number) => void }) {
-  const raw = useStore((s) => s.transform?.[2])
-  const zoom = Number.isFinite(raw) && (raw as number) > 0 ? (raw as number) : 1
-  const prev = useRef(zoom)
-  const rafRef = useRef<number | null>(null)
-  const pendingZoomRef = useRef<number>(zoom)
-  useEffect(() => {
-    if (Math.abs(prev.current - zoom) < 1e-9) return
-    prev.current = zoom
-    pendingZoomRef.current = zoom
-    if (rafRef.current != null) return
-    rafRef.current = requestAnimationFrame(() => {
-      rafRef.current = null
-      onMapZoomChange(pendingZoomRef.current)
-    })
-  }, [zoom, onMapZoomChange])
-  useEffect(() => {
-    return () => {
-      if (rafRef.current != null) cancelAnimationFrame(rafRef.current)
-    }
-  }, [])
-  return null
-}
-
-/**
- * Registers setZoom(z) so the header slider can set viewport zoom while keeping the view center fixed.
- */
-function SliderZoomControl({
-  onMapZoomChange,
-  onSetZoomReady,
-}: {
-  onMapZoomChange: (z: number) => void
-  onSetZoomReady: (setZoom: (z: number) => void) => void
-}) {
-  const { getViewport, setViewport } = useReactFlow()
-  const storeApi = useStoreApi()
-  useEffect(() => {
-    const setZoom = (targetZoom: number) => {
-      const z = Math.min(40, Math.max(0.2, Number(targetZoom) || 0.2))
-      const apply = () => {
-        const domNode = storeApi.getState().domNode
-        if (!domNode || domNode.getBoundingClientRect().width <= 0) return false
-        const vp = getViewport()
-        const rect = domNode.getBoundingClientRect()
-        const w = Math.max(rect.width, 1)
-        const h = Math.max(rect.height, 1)
-        const vz = Math.max(Number(vp.zoom) || 0.2, 0.2)
-        const vx = Number.isFinite(vp.x) ? vp.x : 0
-        const vy = Number.isFinite(vp.y) ? vp.y : 0
-        const cx = (w / 2 - vx) / vz
-        const cy = (h / 2 - vy) / vz
-        const nx = w / 2 - cx * z
-        const ny = h / 2 - cy * z
-        if (!Number.isFinite(nx) || !Number.isFinite(ny)) return false
-        setViewport({ x: nx, y: ny, zoom: z })
-        onMapZoomChange(z)
-        return true
-      }
-      if (apply()) return
-      let n = 0
-      const tick = () => {
-        if (apply()) return
-        if (++n >= 30) return
-        requestAnimationFrame(tick)
-      }
-      requestAnimationFrame(tick)
-    }
-    onSetZoomReady(setZoom)
-  }, [getViewport, setViewport, storeApi, onMapZoomChange, onSetZoomReady])
-  return null
+  stellarCartography?: StellarCartographyMapUi
 }
 
 /** Max time to wait for initial viewport fit before showing the map anyway (avoids staying invisible if fit never runs). */
 const INITIAL_FIT_REVEAL_MS = 250
+
+const DEFAULT_WORMHOLE_DISPLAY_MODE: WormholeDisplayMode = 'always'
 
 export function MapGraph({
   data,
@@ -806,71 +65,9 @@ export function MapGraph({
   onMapZoomChange,
   onSetZoomReady,
   planetLabelOptions = DEFAULT_PLANET_LABEL_OPTIONS,
-  analyticScope = null,
-  stellarCartographySampleEnabled = false,
-  cartographyLayerVisibility,
-  cartographySettingsGates,
-  wormholeDisplayMode = 'always',
+  stellarCartography,
 }: MapGraphProps) {
   const [initialFitDone, setInitialFitDone] = useState(false)
-  const [wormholeHoverLines, setWormholeHoverLines] = useState<string[] | null>(null)
-  const [wormholeRecenterPulseTarget, setWormholeRecenterPulseTarget] =
-    useState<WormholeRecenterPulseTarget | null>(null)
-  const [wormholeLineRevealKey, setWormholeLineRevealKey] = useState<string | null>(null)
-  const wormholeLineRevealClearRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-  const [planetLabelHoverActive, setPlanetLabelHoverActive] = useState(false)
-
-  const wormholeLineReveal = useMemo<WormholeLineRevealApi>(
-    () => ({
-      revealAt: (mapX, mapY) => {
-        if (wormholeLineRevealClearRef.current != null) {
-          clearTimeout(wormholeLineRevealClearRef.current)
-          wormholeLineRevealClearRef.current = null
-        }
-        setWormholeLineRevealKey(wormholeMapCellKey(mapX, mapY))
-      },
-      scheduleClear: () => {
-        if (wormholeLineRevealClearRef.current != null) {
-          clearTimeout(wormholeLineRevealClearRef.current)
-        }
-        wormholeLineRevealClearRef.current = setTimeout(() => {
-          wormholeLineRevealClearRef.current = null
-          setWormholeLineRevealKey(null)
-        }, WORMHOLE_LINE_REVEAL_CLEAR_MS)
-      },
-      cancelClear: () => {
-        if (wormholeLineRevealClearRef.current != null) {
-          clearTimeout(wormholeLineRevealClearRef.current)
-          wormholeLineRevealClearRef.current = null
-        }
-      },
-    }),
-    []
-  )
-
-  useEffect(() => {
-    return () => {
-      if (wormholeLineRevealClearRef.current != null) {
-        clearTimeout(wormholeLineRevealClearRef.current)
-      }
-    }
-  }, [])
-
-  const pulseWormholeAt = useCallback((mapX: number, mapY: number) => {
-    setWormholeRecenterPulseTarget({ mapX, mapY, token: Date.now() })
-  }, [])
-
-  useEffect(() => {
-    if (wormholeRecenterPulseTarget == null) return
-    const t = setTimeout(() => setWormholeRecenterPulseTarget(null), WORMHOLE_RECENTER_PULSE_MS)
-    return () => clearTimeout(t)
-  }, [wormholeRecenterPulseTarget])
-
-  const onPlanetLabelHoverActiveChange = useCallback((active: boolean) => {
-    setPlanetLabelHoverActive(active)
-    if (active) setWormholeHoverLines(null)
-  }, [])
-
   const onInitialFitDone = useCallback(() => setInitialFitDone(true), [])
 
   useEffect(() => {
@@ -879,16 +76,6 @@ export function MapGraph({
   }, [])
 
   const nodes = useMemo(() => toFlowNodes(data.nodes), [data.nodes])
-  const visibleMapEdges = useMemo(
-    () =>
-      filterWormholeEdgesForDisplayMode(
-        data.edges,
-        wormholeDisplayMode,
-        wormholeLineRevealKey
-      ),
-    [data.edges, wormholeDisplayMode, wormholeLineRevealKey]
-  )
-  const edges = useMemo(() => toEdges(visibleMapEdges), [visibleMapEdges])
   const planetMapNodes = useMemo(
     () => data.nodes.filter((n) => n.planet != null),
     [data.nodes]
@@ -912,6 +99,9 @@ export function MapGraph({
     [data.edges, data.wormholeUnknownEntrances]
   )
 
+  const wormholeDisplayMode =
+    stellarCartography?.wormholeDisplayMode ?? DEFAULT_WORMHOLE_DISPLAY_MODE
+
   return (
     <div
       className={`map-graph-cursor-default relative min-h-0 overflow-hidden bg-black ${className ?? 'h-[320px] w-full min-w-0'}`}
@@ -920,70 +110,138 @@ export function MapGraph({
         className="h-full w-full transition-opacity duration-150"
         style={{ opacity: initialFitDone ? 1 : 0 }}
       >
-        <ReactFlow
-          nodes={nodes}
-          edges={edges}
-          nodeTypes={nodeTypes}
-          edgeTypes={edgeTypes}
-          defaultViewport={{ x: 0, y: 0, zoom: 1 }}
-          fitView={false}
-          minZoom={0.2}
-          maxZoom={40}
-          proOptions={{ hideAttribution: true }}
-          nodesDraggable={false}
-          nodesConnectable={false}
-          elementsSelectable={false}
-          panOnDrag
-          zoomOnScroll
-          zoomOnPinch
-        >
-          <WormholeHoverContext.Provider value={setWormholeHoverLines}>
-            <WormholeLineRevealContext.Provider value={wormholeLineReveal}>
-            <WormholeRecenterPulseContext.Provider value={pulseWormholeAt}>
-            <InitialViewportFit
-              nodes={data.nodes}
-              onInitialFitDone={onInitialFitDone}
-              onMapZoomChange={onMapZoomChange}
-            />
-            <ViewportZoomSync onMapZoomChange={onMapZoomChange} />
-            <SliderZoomControl onMapZoomChange={onMapZoomChange} onSetZoomReady={onSetZoomReady} />
-            <CoordinateGridOverlay />
-            <StellarCartographyOverlayPane
-              overlayCircles={data.overlayCircles}
-              wormholeEndpoints={wormholeEndpoints}
-              wormholeEndpointHoverByCell={wormholeEndpointHoverByCell}
-              wormholeRecenterPulseTarget={wormholeRecenterPulseTarget}
-              blockedByPlanetHover={planetLabelHoverActive}
-              nuIonStorms={data.nuIonStorms}
-            />
-            <NormalWarpWellOutlinesOverlay mapNodes={planetMapNodes} />
-            <FixedSizeDotsOverlay
-              planetGrid={planetGrid}
-              planetLabelOptions={planetLabelOptions}
-              labelSourceByNodeId={labelSourceByNodeId}
-              mapNodes={planetMapNodes}
-              routeWaypoints={data.routeWaypoints}
-              waypointGrid={waypointGrid}
-              onPlanetLabelHoverActiveChange={onPlanetLabelHoverActiveChange}
-            />
-            <FlowCoordinateReadout />
-            {cartographyLayerVisibility != null && cartographySettingsGates != null ? (
-              <StellarCartographyHoverPanel
-                analyticScope={analyticScope}
-                sampleEnabled={stellarCartographySampleEnabled}
-                layerVisibility={cartographyLayerVisibility}
-                settingsGates={cartographySettingsGates}
-                wormholeDisplayMode={wormholeDisplayMode}
-                wormholeHoverLines={wormholeHoverLines}
-                blockedByPlanetHover={planetLabelHoverActive}
-                clientToFlowPosition={clientToFlowPosition}
-              />
-            ) : null}
-            </WormholeRecenterPulseContext.Provider>
-            </WormholeLineRevealContext.Provider>
-          </WormholeHoverContext.Provider>
-        </ReactFlow>
+        <WormholeInteractionProvider>
+          <MapGraphFlow
+            data={data}
+            nodes={nodes}
+            planetMapNodes={planetMapNodes}
+            planetGrid={planetGrid}
+            waypointGrid={waypointGrid}
+            labelSourceByNodeId={labelSourceByNodeId}
+            wormholeEndpoints={wormholeEndpoints}
+            wormholeEndpointHoverByCell={wormholeEndpointHoverByCell}
+            wormholeDisplayMode={wormholeDisplayMode}
+            planetLabelOptions={planetLabelOptions}
+            stellarCartography={stellarCartography}
+            onMapZoomChange={onMapZoomChange}
+            onSetZoomReady={onSetZoomReady}
+            onInitialFitDone={onInitialFitDone}
+          />
+        </WormholeInteractionProvider>
       </div>
     </div>
+  )
+}
+
+type MapGraphFlowProps = {
+  data: CombinedMapData
+  nodes: ReturnType<typeof toFlowNodes>
+  planetMapNodes: CombinedMapData['nodes']
+  planetGrid: ReturnType<typeof buildPlanetSpatialGrid>
+  waypointGrid: ReturnType<typeof buildPlanetSpatialGrid> | null
+  labelSourceByNodeId: ReturnType<typeof buildLabelSourceByNodeId>
+  wormholeEndpoints: { x: number; y: number }[]
+  wormholeEndpointHoverByCell: ReturnType<typeof buildWormholeEndpointHoverIndex>
+  wormholeDisplayMode: WormholeDisplayMode
+  planetLabelOptions: PlanetLabelOptions
+  stellarCartography?: StellarCartographyMapUi
+  onMapZoomChange: (zoom: number) => void
+  onSetZoomReady: (setZoom: (zoom: number) => void) => void
+  onInitialFitDone: () => void
+}
+
+function MapGraphFlow({
+  data,
+  nodes,
+  planetMapNodes,
+  planetGrid,
+  waypointGrid,
+  labelSourceByNodeId,
+  wormholeEndpoints,
+  wormholeEndpointHoverByCell,
+  wormholeDisplayMode,
+  planetLabelOptions,
+  stellarCartography,
+  onMapZoomChange,
+  onSetZoomReady,
+  onInitialFitDone,
+}: MapGraphFlowProps) {
+  const {
+    wormholeLineRevealKey,
+    wormholeHoverLines,
+    wormholeRecenterPulseTarget,
+    blockedByPlanetHover,
+    onPlanetLabelHoverActiveChange,
+  } = useWormholeInteractionState()
+
+  const visibleMapEdges = useMemo(
+    () =>
+      filterWormholeEdgesForDisplayMode(
+        data.edges,
+        wormholeDisplayMode,
+        wormholeLineRevealKey
+      ),
+    [data.edges, wormholeDisplayMode, wormholeLineRevealKey]
+  )
+  const edges = useMemo(() => toEdges(visibleMapEdges), [visibleMapEdges])
+
+  return (
+    <ReactFlow
+      nodes={nodes}
+      edges={edges}
+      nodeTypes={nodeTypes}
+      edgeTypes={edgeTypes}
+      defaultViewport={{ x: 0, y: 0, zoom: 1 }}
+      fitView={false}
+      minZoom={0.2}
+      maxZoom={40}
+      proOptions={{ hideAttribution: true }}
+      nodesDraggable={false}
+      nodesConnectable={false}
+      elementsSelectable={false}
+      panOnDrag
+      zoomOnScroll
+      zoomOnPinch
+    >
+      <InitialViewportFit
+        nodes={data.nodes}
+        onInitialFitDone={onInitialFitDone}
+        onMapZoomChange={onMapZoomChange}
+      />
+      <ViewportZoomSync onMapZoomChange={onMapZoomChange} />
+      <SliderZoomControl onMapZoomChange={onMapZoomChange} onSetZoomReady={onSetZoomReady} />
+      <CoordinateGridOverlay />
+      <StellarCartographyOverlayPane
+        overlayCircles={data.overlayCircles}
+        wormholeEndpoints={wormholeEndpoints}
+        wormholeEndpointHoverByCell={wormholeEndpointHoverByCell}
+        wormholeRecenterPulseTarget={wormholeRecenterPulseTarget}
+        blockedByPlanetHover={blockedByPlanetHover}
+        nuIonStorms={data.nuIonStorms}
+      />
+      <NormalWarpWellOutlinesOverlay mapNodes={planetMapNodes} />
+      <FixedSizeDotsOverlay
+        planetGrid={planetGrid}
+        planetLabelOptions={planetLabelOptions}
+        labelSourceByNodeId={labelSourceByNodeId}
+        mapNodes={planetMapNodes}
+        routeWaypoints={data.routeWaypoints}
+        waypointGrid={waypointGrid}
+        onPlanetLabelHoverActiveChange={onPlanetLabelHoverActiveChange}
+      />
+      <FlowCoordinateReadout />
+      {stellarCartography != null ? (
+        <StellarCartographyHoverPanel
+          analyticScope={stellarCartography.analyticScope}
+          sampleEnabled={stellarCartography.sampleEnabled}
+          layerVisibility={stellarCartography.layerVisibility}
+          settingsGates={stellarCartography.settingsGates}
+          wormholeDisplayMode={stellarCartography.wormholeDisplayMode}
+          wormholeHoverLines={wormholeHoverLines}
+          blockedByPlanetHover={blockedByPlanetHover}
+          clientToFlowPosition={clientToFlowPosition}
+        />
+      ) : null}
+    </ReactFlow>
   )
 }
