+++ packages/frontend/src/components/map-graph/viewportControls.tsx
@@ -1,0 +1,161 @@
+import { useCallback, useEffect, useRef, useState } from 'react'
+import { useReactFlow, useStore, useStoreApi } from '@xyflow/react'
+import type { CombinedMapData } from '../../api/bff'
+import {
+  CELL_CENTER_OFFSET,
+  gameMapYToFlowCenterY,
+} from './geometry'
+
+/** Fraction of display area to leave as blank margin on each side when fitting initial view (0.1 = 10%). */
+const INITIAL_FIT_MARGIN = 0.1
+
+/**
+ * Computes initial viewport so the bounding rectangle of all node centers is
+ * centered with a 10% margin (whichever dimension is most constrained).
+ * Runs once when the pane has size and nodes are present. Calls onInitialFitDone when done (or when no fit will run).
+ */
+export function InitialViewportFit({
+  nodes,
+  onInitialFitDone,
+  onMapZoomChange,
+}: {
+  nodes: CombinedMapData['nodes']
+  onInitialFitDone: () => void
+  onMapZoomChange: (zoom: number) => void
+}) {
+  const { setViewport } = useReactFlow()
+  const domNode = useStore((s) => s.domNode ?? null)
+  const [size, setSize] = useState({ width: 0, height: 0 })
+  const hasFittedRef = useRef(false)
+  const doneCalledRef = useRef(false)
+
+  const callDoneOnce = useCallback(() => {
+    if (doneCalledRef.current) return
+    doneCalledRef.current = true
+    onInitialFitDone()
+  }, [onInitialFitDone])
+
+  useEffect(() => {
+    if (!domNode) return
+    let raf = 0
+    const ro = new ResizeObserver((entries) => {
+      const { width, height } = entries[0]?.contentRect ?? { width: 0, height: 0 }
+      cancelAnimationFrame(raf)
+      raf = requestAnimationFrame(() => setSize({ width, height }))
+    })
+    ro.observe(domNode)
+    return () => {
+      cancelAnimationFrame(raf)
+      ro.disconnect()
+    }
+  }, [domNode])
+
+  useEffect(() => {
+    if (nodes.length === 0) {
+      callDoneOnce()
+      return
+    }
+    if (size.width <= 0 || size.height <= 0 || hasFittedRef.current) return
+    const xs = nodes.map((n) => Number(n.x)).filter(Number.isFinite)
+    const ys = nodes.map((n) => Number(n.y)).filter(Number.isFinite)
+    if (xs.length === 0 || ys.length === 0) {
+      callDoneOnce()
+      return
+    }
+    const minFx = Math.min(...xs) + CELL_CENTER_OFFSET
+    const maxFx = Math.max(...xs) + CELL_CENTER_OFFSET
+    const flowCentersY = ys.map((py) => gameMapYToFlowCenterY(py))
+    const minFy = Math.min(...flowCentersY)
+    const maxFy = Math.max(...flowCentersY)
+    const contentWidth = Math.max(maxFx - minFx, 1)
+    const contentHeight = Math.max(maxFy - minFy, 1)
+    const centerX = (minFx + maxFx) / 2
+    const centerY = (minFy + maxFy) / 2
+    const usableW = size.width * (1 - 2 * INITIAL_FIT_MARGIN)
+    const usableH = size.height * (1 - 2 * INITIAL_FIT_MARGIN)
+    const scaleW = usableW / contentWidth
+    const scaleH = usableH / contentHeight
+    const zoom = Math.min(40, Math.max(0.2, Math.min(scaleW, scaleH)))
+    const x = size.width / 2 - centerX * zoom
+    const y = size.height / 2 - centerY * zoom
+    hasFittedRef.current = true
+    setViewport({ x, y, zoom })
+    onMapZoomChange(zoom)
+    callDoneOnce()
+  }, [nodes, size, setViewport, callDoneOnce, onMapZoomChange])
+
+  return null
+}
+
+/** Mirrors React Flow zoom to the app (wheel, pinch, initial fit, slider). */
+export function ViewportZoomSync({ onMapZoomChange }: { onMapZoomChange: (z: number) => void }) {
+  const raw = useStore((s) => s.transform?.[2])
+  const zoom = Number.isFinite(raw) && (raw as number) > 0 ? (raw as number) : 1
+  const prev = useRef(zoom)
+  const rafRef = useRef<number | null>(null)
+  const pendingZoomRef = useRef<number>(zoom)
+  useEffect(() => {
+    if (Math.abs(prev.current - zoom) < 1e-9) return
+    prev.current = zoom
+    pendingZoomRef.current = zoom
+    if (rafRef.current != null) return
+    rafRef.current = requestAnimationFrame(() => {
+      rafRef.current = null
+      onMapZoomChange(pendingZoomRef.current)
+    })
+  }, [zoom, onMapZoomChange])
+  useEffect(() => {
+    return () => {
+      if (rafRef.current != null) cancelAnimationFrame(rafRef.current)
+    }
+  }, [])
+  return null
+}
+
+/**
+ * Registers setZoom(z) so the header slider can set viewport zoom while keeping the view center fixed.
+ */
+export function SliderZoomControl({
+  onMapZoomChange,
+  onSetZoomReady,
+}: {
+  onMapZoomChange: (z: number) => void
+  onSetZoomReady: (setZoom: (z: number) => void) => void
+}) {
+  const { getViewport, setViewport } = useReactFlow()
+  const storeApi = useStoreApi()
+  useEffect(() => {
+    const setZoom = (targetZoom: number) => {
+      const z = Math.min(40, Math.max(0.2, Number(targetZoom) || 0.2))
+      const apply = () => {
+        const domNode = storeApi.getState().domNode
+        if (!domNode || domNode.getBoundingClientRect().width <= 0) return false
+        const vp = getViewport()
+        const rect = domNode.getBoundingClientRect()
+        const w = Math.max(rect.width, 1)
+        const h = Math.max(rect.height, 1)
+        const vz = Math.max(Number(vp.zoom) || 0.2, 0.2)
+        const vx = Number.isFinite(vp.x) ? vp.x : 0
+        const vy = Number.isFinite(vp.y) ? vp.y : 0
+        const cx = (w / 2 - vx) / vz
+        const cy = (h / 2 - vy) / vz
+        const nx = w / 2 - cx * z
+        const ny = h / 2 - cy * z
+        if (!Number.isFinite(nx) || !Number.isFinite(ny)) return false
+        setViewport({ x: nx, y: ny, zoom: z })
+        onMapZoomChange(z)
+        return true
+      }
+      if (apply()) return
+      let n = 0
+      const tick = () => {
+        if (apply()) return
+        if (++n >= 30) return
+        requestAnimationFrame(tick)
+      }
+      requestAnimationFrame(tick)
+    }
+    onSetZoomReady(setZoom)
+  }, [getViewport, setViewport, storeApi, onMapZoomChange, onSetZoomReady])
+  return null
+}
